refactor: simplify namespaces
