--- dragonfly_doe2/doe/__init__.py
@@ -1,1 +1,0 @@
-"""Doe2 components sub-module."""
