--- dragonfly_doe2/doe/geometry.py
@@ -1,3 +1,0 @@
-"""Some cls for LBT to DOE2 format to keep properties classes clean"""
-
-
